move customCharacter module inside player module
