+++ src/__tests__/player/modules/playerCommon.module.ts
@@ -4,11 +4,16 @@ import {mongooseOptions, mongoString} from "../../test_utils/const/db";
 import {ModelName} from "../../../common/enum/modelName.enum";
 import {RequestHelperModule} from "../../../requestHelper/requestHelper.module";
 import {PlayerSchema} from "../../../player/player.schema";
-import {CustomCharacterModule} from "../../../customCharacter/customCharacter.module";
 import {PlayerService} from "../../../player/player.service";
 import {isPlayerExists} from "../../../player/decorator/validation/IsPlayerExists.decorator";
-import {ClanInventoryModule} from "../../../clanInventory/clanInventory.module";
-import {ClanModule} from "../../../clan/clan.module";
+import {CustomCharacterSchema} from "../../../player/customCharacter/customCharacter.schema";
+import {AuthorizationModule} from "../../../authorization/authorization.module";
+import {CustomCharacterService} from "../../../player/customCharacter/customCharacter.service";
+import {
+    isCustomCharacterExists
+} from "../../../player/customCharacter/decorator/validation/IsCustomCharacterExists.decorator";
+import {ClanSchema} from "../../../clan/clan.schema";
+import {RoomSchema} from "../../../clanInventory/room/room.schema";
 
 export default class PlayerCommonModule {
     private constructor() {
@@ -22,15 +27,17 @@ export default class PlayerCommonModule {
                 imports: [
                     MongooseModule.forRoot(mongoString, mongooseOptions),
                     MongooseModule.forFeature([
-                        {name: ModelName.PLAYER, schema: PlayerSchema}
+                        {name: ModelName.PLAYER, schema: PlayerSchema},
+                        {name: ModelName.CLAN, schema: ClanSchema},
+                        {name: ModelName.ROOM, schema: RoomSchema},
+                        {name: ModelName.CUSTOM_CHARACTER, schema: CustomCharacterSchema}
                     ]),
-                    CustomCharacterModule,
                     RequestHelperModule,
-                    ClanInventoryModule,
-                    ClanModule
+                    AuthorizationModule
                 ],
                 providers: [
-                    PlayerService, isPlayerExists
+                    PlayerService, isPlayerExists,
+                    CustomCharacterService, isCustomCharacterExists
                 ],
             }).compile();
