+++ src/authorization/caslAbility.factory.ts
@@ -7,8 +7,8 @@ import { ProfileDto } from "../profile/dto/profile.dto";
 import { PlayerDto } from "../player/dto/player.dto";
 import { UpdateProfileDto } from "../profile/dto/updateProfile.dto";
 import { UpdatePlayerDto } from "../player/dto/updatePlayer.dto";
-import { CustomCharacterDto } from "../customCharacter/dto/customCharacter.dto";
-import { UpdateCustomCharacterDto } from "../customCharacter/dto/updateCustomCharacter.dto";
+import { CustomCharacterDto } from "../player/customCharacter/dto/customCharacter.dto";
+import { UpdateCustomCharacterDto } from "../player/customCharacter/dto/updateCustomCharacter.dto";
 import { profileRules } from "./rule/profileRules";
 import { playerRules } from "./rule/playerRules";
 import { customCharacterRules } from "./rule/customCharacterRules";
