--- src/characterClass/MikaR.txt
@@ -1,1 +1,0 @@
-Hello world
