+++ src/characterClass/dto/characterClass.dto.ts
@@ -1,7 +1,7 @@
 import {Expose, Type} from "class-transformer";
 import {ExtractField} from "../../common/decorator/response/ExtractField";
 import {GestaltCycle} from "../../common/enum/gestaltCycle.enum";
-import {CustomCharacterDto} from "../../customCharacter/dto/customCharacter.dto";
+import {CustomCharacterDto} from "../../player/customCharacter/dto/customCharacter.dto";
 import AddType from "../../common/base/decorator/AddType.decorator";
 
 /**
