--- src/characterClass/files/alexei.txt
@@ -1,1 +1,0 @@
-Hello world
