--- src/characterClass/files/niklas.txt
@@ -1,1 +1,0 @@
-"Hello world"
