+++ src/player/customCharacter/const/CharacterBaseStats.ts
