+++ src/player/customCharacter/enum/characterId.enum.ts
