+++ src/player/dto/player.dto.ts
@@ -1,7 +1,7 @@
 import {Expose, Type} from "class-transformer";
 import {ClanDto} from "../../clan/dto/clan.dto";
 import {ExtractField} from "../../common/decorator/response/ExtractField";
-import {CustomCharacterDto} from "../../customCharacter/dto/customCharacter.dto";
+import {CustomCharacterDto} from "../customCharacter/dto/customCharacter.dto";
 import AddType from "../../common/base/decorator/AddType.decorator";
 import { GameStatisticsDto } from "./gameStatistics.dto";
 import {ArrayMaxSize, IsArray, IsInt, IsMongoId, IsOptional} from "class-validator";
